I have refactored the services to use a generic CRUDService, which simplifies the implementation and aligns with the user's request for consistency.
