+++ backend/app/services/consignment/statement_service.py
@@ -11,7 +11,8 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy import and_, func, select
 from sqlalchemy.orm import selectinload
 
-from ..base import CRUDService, transactional, audit_log
+from ..generics import CRUDService
+from ..base import transactional, audit_log
 from ..exceptions import ValidationError, BusinessRuleError, ConsignmentError
 from ...models import (
     ConsignmentStatement, ConsignmentAgreement, Consignment,
