+++ backend/app/services/contract/reservation_service.py
@@ -10,7 +10,8 @@ from datetime import datetime, date
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy import and_, func, select
 
-from ..base import CRUDService, transactional, audit_log
+from ..generics import CRUDService
+from ..base import transactional, audit_log
 from ..exceptions import ValidationError, BusinessRuleError, ContractError, NotFoundError
 from ...models import ContractReservation, TenderContract, Product, Batch, Allocation
 from ...schemas import ContractReservationSchema, ContractReservationCreateSchema, ContractReservationUpdateSchema
