+++ backend/app/services/customer/customer_service.py
@@ -9,7 +9,8 @@ from typing import Dict, Any, List, Optional
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy import and_, select
 
-from ..base import CRUDService, transactional, audit_log
+from ..generics import CRUDService
+from ..base import transactional, audit_log
 from ..exceptions import ValidationError, ConflictError, NotFoundError
 from ...models import Customer, CustomerType, SectorType, CustomerAddress
 from ...schemas import CustomerSchema, CustomerCreateSchema, CustomerUpdateSchema
