+++ backend/app/services/product/batch_service.py
@@ -10,7 +10,8 @@ from datetime import datetime, date,timedelta
 from sqlalchemy.orm import Session
 from sqlalchemy import and_, or_, select, func
 
-from ..base import CRUDService, transactional, audit_log
+from ..generics import CRUDService
+from ..base import transactional, audit_log
 from ..exceptions import ValidationError, BusinessRuleError, NotFoundError
 from ...models import Batch, Product, Allocation
 from ...schemas import BatchSchema, BatchCreateSchema, BatchUpdateSchema
