+++ backend/app/services/sales/packing_slip_service.py
@@ -10,7 +10,8 @@ from datetime import datetime, date
 from sqlalchemy.orm import Session
 from sqlalchemy import and_, func
 
-from ..base import CRUDService, transactional, audit_log
+from ..generics import CRUDService
+from ..base import transactional, audit_log
 from ..exceptions import ValidationError, BusinessRuleError, NotFoundError
 from ...models import PackingSlip, SalesOrder, Customer
 from ...schemas import PackingSlipSchema, PackingSlipCreateSchema, PackingSlipUpdateSchema
