+++ backend/app/services/shipping/shipment_service.py
@@ -10,7 +10,8 @@ from datetime import datetime, date, timedelta
 from sqlalchemy.orm import Session
 from sqlalchemy import and_, func
 
-from ..base import CRUDService, transactional, audit_log
+from ..generics import CRUDService
+from ..base import transactional, audit_log
 from ..exceptions import ValidationError, BusinessRuleError, ShipmentError, NotFoundError
 from ...models import (
     Shipment, PackingSlip, PackingOrder, Customer, Carrier,
