+++ backend/app/services/shipping/shipping_method_service.py
@@ -1,5 +1,5 @@
 from sqlalchemy.orm import Session
-from ..base import CRUDService
+from ..generics import CRUDService
 from ...models.helper import ShippingMethod
 from ...schemas.shipping_method import ShippingMethodCreateSchema, ShippingMethodUpdateSchema, ShippingMethodSchema
