+++ backend/app/services/system/notification_type_service.py
@@ -1,5 +1,5 @@
 from sqlalchemy.orm import Session
-from ..base import CRUDService
+from ..generics import CRUDService
 from ...models.helper import NotificationType
 from ...schemas.notification_type import NotificationTypeCreateSchema, NotificationTypeUpdateSchema, NotificationTypeSchema
