+++ backend/app/services/warehouse_ops/packing_service.py
@@ -10,7 +10,8 @@ from datetime import datetime, date
 from sqlalchemy.orm import Session
 from sqlalchemy import and_, func
 
-from ..base import CRUDService, transactional, audit_log
+from ..generics import CRUDService
+from ..base import transactional, audit_log
 from ..exceptions import ValidationError, BusinessRuleError, PackingError, NotFoundError
 from ...models import (
     PackingOrder, PackingBox, PackingBoxItem, PickingList,
